Refactor polygon center support into polyline base
